+++ pandas2pgfplots/__init__.py
@@ -1,4 +1,4 @@
 from pandas2pgfplots.plots import scatter, cactus, Plot
 from pandas2pgfplots.utils import tikzify_dict, deep_update
 
-__version__ = "0.1.0"
+__version__ = "0.1.0dev"