+++ CHANGELOG.md
@@ -4,6 +4,14 @@ The project provides a basic functionality to produce [pgfplots] code from
 [pandas] DataFrames. 
 
 ## [Unreleased]
+### Added
+ * `display` module that can display plots as SVG in Jupyter
+ * wrapper class `Plot` that renders plots in Jupyter as SVG (and pgfplots code
+     as string in Python console)
+
+### Changed
+ * `tikzify` changed to `tikzify_dict`
+ * functions producing plots return a Plot object instead of string
 
 ## [0.1.0] — 2020-05-08
 ### Added
@@ -11,8 +19,6 @@ The project provides a basic functionality to produce [pgfplots] code from
    that can be passed as tikz options 
  * scatter plot (function `scatter`)
  * cactus plot (function `cactus`)
-
-
 
 [Unreleased]: https://github.com/xblahoud/pandas2pgfplots/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/xblahoud/pandas2pgfplots/tree/v0.1.0